Feature/38 get client companylist

+++ src/main/java/com/fiveguys/RIA/RIA_Backend/user/model/dto/response/MyProjectResponseDto.java
@@ -1,5 +1,6 @@
 package com.fiveguys.RIA.RIA_Backend.user.model.dto.response;
 
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -13,8 +14,8 @@ public class MyProjectResponseDto {
   private String clientCompanyName;
   private LocalDate startDay;
   private LocalDate endDay;
-  private Integer expectedRevenue;
-  private Double expectedMarginRate;
+  private BigDecimal expectedRevenue;
+  private BigDecimal expectedMarginRate;
   private Integer expectedProfit;
   private String salesManagerName;
   private String status;
